Create machine_learning
This project applies random forest regression to predict a time series. Cielo's stock prices were used as a practical example for this modeling. 
All machine learning workflow was developed and explained. The time series problem switched it to supervised learning. 
The code was written in R, considering the tidymodels structure.